+++ lib/get_state_manager/get_state_manager.dart
@@ -1,4 +1,3 @@
-export 'src/rx_flutter/rx_getx_widget.dart';
 export 'src/rx_flutter/rx_notifier.dart';
 export 'src/rx_flutter/rx_obx_widget.dart';
 export 'src/rx_flutter/rx_ticker_provider_mixin.dart';